+++ src/store/assets/ibc/getters.js
@@ -1,0 +1,6 @@
+export default {
+  error: (state) => state.error,
+  isLoading: (state) => state.isLoading,
+  isHandling: (state) => state.isHandling,
+  modal: (state) => state.modal,
+};