+++ src/views/assets/AssetsViewDetailInfo.vue
@@ -43,7 +43,7 @@ export default {
     AssetsViewDetailMarketing,
   },
   computed: {
-    ...mapGetters('assets', ['error', 'isLoading']),
+    ...mapGetters('assetsDetail', ['error', 'isLoading']),
     address() {
       return this.$route.params.id;
     },
@@ -52,7 +52,7 @@ export default {
     this.initAssetsDetail(this.address);
   },
   methods: {
-    ...mapActions('assets', ['initAssetsDetail']),
+    ...mapActions('assetsDetail', ['initAssetsDetail']),
   },
 };
 </script>