+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [develop] 4.11.0 - 2024-03-29
+
+### Added
+
+- Add assets IBC transfer
+
 ## [4.10.11] - 2024-03-05
 
 ### Added